requirements.txt
numpy>=1.21.0
Pillow>=8.3.1
opencv-python>=4.5.3
matplotlib>=3.4.3